requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
djoser==2.3.3
python-decouple==3.8
django-cors-headers==4.7.0